Improve performance by quickly checking first for any cycle
The job did not finish on the Angular CLI (which is quite large), but with this upgrade
it finishes in a little over 10 seconds on my machine.

Also we only check for cycles once per file (instead of checking for every import).

The tests had to be changed because the order from TSLint seems to have changed
with the new algorithm. case1.1 is ran after case1 and so will not detect the
cycle before case1.1 (and will report case1.1 first). Same goes for the other
tests.